Split up prototype into multiple files; resolve #23
