+++ analyzer/system/__init__.py
@@ -1,0 +1,2 @@
+from .centos import *
+from .ubuntu import *
